Solution to BUG when clicking on highlighted area inside item
I was considering clicking "inside an item" as equivalent to clicking inside an
anchor inside the menu. But there was a case that I forgot: when someone clicks
in the "highlighted area" of the text inside the anchor. That area is achieved
by surrounding the corresponding text with a <strong> tag. So, when clicking on
that area, the target for the event fired will not be the anchor, but the strong
tag. This made it necessary to check not only if the tagName for the target is
equal to 'A', but also if the tagName of the parent of the target is (this is
the case when we click inside the highlighted area).